requirements.txt
python-telegram-bot
simpleeval
